app/requirements.txt
aiohttp
celery[redis]
python-decouple
opencv-python